+++ src/db/repositories/user-episode.ts
@@ -2,10 +2,14 @@ import type { InsertUserEpisode } from '@/domain/entities/user-episode'
 import { db } from '..'
 import { schema } from '../schema'
 import type { GetUserEpisodesInput } from '@/domain/services/user-episodes/get-user-episodes'
-import { and, eq, inArray } from 'drizzle-orm'
+import { and, count, desc, eq, inArray } from 'drizzle-orm'
 
 export async function insertUserEpisodes(values: InsertUserEpisode[]) {
-  return db.insert(schema.userEpisodes).values(values).returning()
+  return db
+    .insert(schema.userEpisodes)
+    .values(values)
+    .returning()
+    .onConflictDoNothing()
 }
 
 export async function selectUserEpisodes({
@@ -18,7 +22,7 @@ export async function selectUserEpisodes({
     .where(
       and(
         eq(schema.userEpisodes.userId, userId),
-        eq(schema.userEpisodes.tmdbId, tmdbId)
+        tmdbId ? eq(schema.userEpisodes.tmdbId, tmdbId) : undefined
       )
     )
     .orderBy(schema.userEpisodes.episodeNumber)
@@ -28,4 +32,14 @@ export async function deleteUserEpisodes(ids: string[]) {
   return db
     .delete(schema.userEpisodes)
     .where(inArray(schema.userEpisodes.id, ids))
+}
+
+export async function selectMostWatched(userId: string) {
+  return db
+    .select({ count: count(), tmdbId: schema.userEpisodes.tmdbId })
+    .from(schema.userEpisodes)
+    .where(and(eq(schema.userEpisodes.userId, userId)))
+    .groupBy(schema.userEpisodes.tmdbId)
+    .orderBy(desc(count()))
+    .limit(3)
 }