+++ src/domain/entities/user-episode.ts
@@ -4,5 +4,5 @@ import type { InferInsertModel, InferSelectModel } from 'drizzle-orm'
 export type UserEpisode = InferSelectModel<typeof schema.userEpisodes>
 export type InsertUserEpisode = Pick<
   InferInsertModel<typeof schema.userEpisodes>,
-  'seasonNumber' | 'episodeNumber' | 'tmdbId' | 'userId'
+  'seasonNumber' | 'episodeNumber' | 'tmdbId' | 'userId' | 'runtime'
 >