+++ src/domain/services/user-items/get-user-item.ts
@@ -1,5 +1,4 @@
 import { selectUserItem } from '@/db/repositories/user-item-repository'
-import { UserItemNotFoundError } from '@/domain/errors/user-item-not-found-error'
 
 export type GetUserItemInput = {
   tmdbId: number