+++ src/http/schemas/user-stats.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod'
 
-export const getUserStatsParamsSchema = z.object({
+export const getUserDefaultSchema = z.object({
   id: z.string(),
 })
 
@@ -10,5 +10,57 @@ export const getUserStatsResponseSchema = {
     followingCount: z.number(),
     watchedMoviesCount: z.number(),
     watchedSeriesCount: z.number(),
+  }),
+}
+
+export const getUserTotalHoursResponseSchema = {
+  200: z.object({
+    totalHours: z.number(),
+  }),
+}
+
+export const getUserReviewsCountResponseSchema = {
+  200: z.object({
+    reviewsCount: z.number(),
+  }),
+}
+
+export const getUserMostWatchedSeriesResponseSchema = {
+  200: z.object({
+    mostWatchedSeries: z.array(
+      z.object({
+        id: z.number(),
+        episodes: z.number(),
+        title: z.string(),
+        posterPath: z.string().nullable(),
+        backdropPath: z.string().nullable(),
+      })
+    ),
+  }),
+}
+
+export const getUserWatchedGenresResponseSchema = {
+  200: z.object({
+    genres: z.array(
+      z.object({
+        name: z.string(),
+        count: z.number(),
+        percentage: z.number(),
+      })
+    ),
+  }),
+}
+
+export const getUserWatchedCastResponseSchema = {
+  200: z.object({
+    watchedCast: z.array(
+      z.object({
+        id: z.string(),
+        name: z.string(),
+        count: z.number(),
+        percentage: z.number(),
+        profilePath: z.string().nullable(),
+      })
+    ),
   }),
 }