+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Add user reviews count and vitest coverage [#47](https://github.com/plotwist-app/plotwist-backend/pull/47)
 - Add simple user stats (watched items total and following/followers total) [#43](https://github.com/plotwist-app/plotwist-backend/pull/43)
 - Add likeCount and userLike in select list by id [#42](https://github.com/plotwist-app/plotwist-backend/pull/42)
 - Add operations for like [#40](https://github.com/plotwist-app/plotwist-backend/pull/40)